refactor: PositionBadge -> 사실상 기본뱃지이므로 Badge로 변경하고 shared로 이동
